Refactor layout and move Clock component to Navigation
In order to streamline the app layout and improve user experience, several changes have been made. The "Clock" component is now part of the "Navigation" component instead of "page.tsx". In addition, unnecessary class names and elements in the "layout.tsx" have been removed. The WorldTimeApi label was also updated in the "page.tsx".