feat: dynamic imports for Client and Clock components
Transitioned to using dynamic imports for the Client and Clock components to enable client-side only rendering. This ensures that these components, which rely on browser features, are correctly rendered and improve the overall functionality of the application. Additionally, this change enhances the app's performance by loading these components only when required.