feat: wtf
Updated the CacheButton to receive a "tag" as a prop and handle revalidation of this tag. Several changes were made to page files (Lol, Lul, and others) including the removal of some components replaced with WorldTimeApis components. The WorldTimeApis components, introduced to fetch different APIs and handle caching options, were also created. Furthermore, the label texts on the Navigation component were updated.